Refactor classes - highlight standalone modules
