+++ executorlib/backend/interactive_serial.py
@@ -2,8 +2,8 @@ import sys
 from os.path import abspath
 from typing import List, Optional
 
-from executorlib.interactive.backend import call_funct, parse_arguments
-from executorlib.shared.communication import (
+from executorlib.standalone.interactive.backend import call_funct, parse_arguments
+from executorlib.standalone.interactive.communication import (
     interface_connect,
     interface_receive,
     interface_send,
