+++ executorlib/interactive/dependencies.py
@@ -2,9 +2,9 @@ from concurrent.futures import Future
 from typing import Any, Callable, Dict
 
 from executorlib.interactive import create_executor
-from executorlib.shared.executor import ExecutorSteps, execute_tasks_with_dependencies
-from executorlib.shared.plot import draw, generate_nodes_and_edges, generate_task_hash
-from executorlib.shared.thread import RaisingThread
+from executorlib.interactive.shared import ExecutorSteps, execute_tasks_with_dependencies
+from executorlib.standalone.plot import draw, generate_nodes_and_edges, generate_task_hash
+from executorlib.standalone.thread import RaisingThread
 
 
 class ExecutorWithDependencies(ExecutorSteps):
