+++ tests/test_executor_backend_mpi_noblock.py
@@ -1,7 +1,7 @@
 import unittest
 
 from executorlib import Executor
-from executorlib.shared.executor import cloudpickle_register
+from executorlib.standalone.serialize import cloudpickle_register
 
 
 def calc(i):
