+++ tests/test_flux_executor.py
@@ -6,7 +6,8 @@ import unittest
 import numpy as np
 
 from executorlib.interactive.executor import InteractiveExecutor
-from executorlib.shared.executor import cloudpickle_register, execute_parallel_tasks
+from executorlib.standalone.serialize import cloudpickle_register
+from executorlib.interactive.shared import execute_parallel_tasks
 
 
 try:
