+++ tests/test_local_executor_future.py
@@ -6,7 +6,7 @@ import unittest
 import numpy as np
 
 from executorlib.interactive.executor import InteractiveExecutor
-from executorlib.shared.spawner import MpiExecSpawner
+from executorlib.standalone.interactive.spawner import MpiExecSpawner
 
 
 skip_mpi4py_test = importlib.util.find_spec("mpi4py") is None
